HTB_BusinessCTF2032/Web/web_polaris_control/challenge/requirements.txt
Flask==2.3.2
neo4j==5.10.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.7.0
cryptography==41.0.1
requests==2.31.0
selenium
Pillow==10.0.0 